requirements.txt
matplotlib==3.5.0
numpy==1.21.4
Pillow==8.4.0
tensorflow==2.7.0
